functions/publish-apply-extension-jobs/requirements.txt
google-cloud-pubsub==0.41.0
google-cloud-firestore==1.2.0
